+++ lib/blocs/Import_order_event.dart
@@ -1,0 +1,10 @@
+part of 'import_order_bloc.dart';
+
+abstract class CovidEvent extends Equatable {
+  const CovidEvent();
+
+  @override
+  List<Object> get props => [];
+}
+
+class GetCovidList extends CovidEvent {}
